audits/gate-never-asked: audit the itoa family for the 097/100 defect class; widen 8 corpora, fix 2 oracles
The same defect class has now been found three times in this repository and twice
in landed code: A GATE THAT NEVER ASKED. The corpus does not generate the input
class that breaks the implementation, so it passes and the defect ships behind a
green tick.

  change 097  a negative `length` is a zero-padded field width, not room.
              Corpus swept Length 0..40. DEFECT: 171600 of 171600 cases.
  change 100  the same, in the 64-bit sibling. DEFECT: 123000 of 123000.
  the itoa family (054/055/056/057/072/073/074/075) -- radix is a SIGNED int and
              all eight corpora swept exactly `for (int radix=2; radix<=36)`.

So the third one was asked. This commit is that audit, its answer, and the
repairs it produced.

NO DEFECT IN ANY OF THE EIGHT IMPLEMENTATIONS.

They are byte-identical to ucrtbase on EVERY out-of-range radix that returns --
37, 100, 255, 256, -1, -2, -10, -36, INT_MAX, INT_MIN -- including the high-bit
garbage characters an invalid radix produces. That is not luck: ucrtbase converts
the radix to a 32-bit UNSIGNED value, so -1 becomes 4294967295 and the division
proceeds normally, and these implementations index the same digit table the same
way. For radix -1 and v=4294967295: 4294967295/4294967295 = 1 remainder 0 -> "10".

WHAT DIFFERS IS CONFINED TO TERMINATION.

  radix 0   ucrtbase 0xC0000094 (integer divide by zero); ours IDENTICAL.
  radix 1   ucrtbase 0xC0000005 (access violation); ours 0xC00000FD (stack
            overflow). This is the `v /= 1` trap that hung change 278's live
            harness: the quotient never decreases, so one implementation walks
            off the buffer and the other off the stack. Both are unbounded loops
            on a caller bug and neither is correct. Nothing is changed for radix
            1 -- there is no right answer to reproduce, only two ways of failing
            to have one.

WHAT THE AUDIT CHANGED

1. ALL EIGHT CORPORA ARE WIDENED to sweep 2..36 AND the ten out-of-range radices
   that return. Radix 0 and 1 are deliberately excluded and each file says why: a
   corpus cannot contain a case that kills the process.

2. TWO REFERENCE MODELS WERE WRONG and the widened corpus caught them on the
   first run:

     FAIL v=4294967295 radix=-1:  ours '10'   sys '10'   ref 'V'
     FAIL v=4294967296 radix=-1:  ours '11'   sys '11'   ref '0'

   OURS MATCHED WINDOWS; THE ORACLE WAS THE ODD ONE OUT. 073-ui64tow and
   075-i64tow converted the radix with (unsigned long long) rather than
   (unsigned), sign-extending -1 to 0xFFFFFFFFFFFFFFFF. For radix 2..36 the two
   conversions are identical, which is exactly why a corpus that stopped at 36
   could never tell them apart; their siblings 055 and 057 already used
   (unsigned). Both fixed. All eight gates pass with the widened corpus.

   The finding was not in the assembly. It was in the thing the assembly is
   checked against -- and only a THREE-WAY gate could see it, because a two-way
   ours-vs-Windows check would have passed silently.

METHOD NOTE: ONE CALL PER PROCESS. The first version of the probe put every call
in one process behind __try and died at exit code 148 having printed nothing --
the signature of a fail-fast, which __except cannot catch. Changes 274 and 276
hit the same wall on SysFreeString. The probe is now one call per process with a
driver that reads the exit code, so a termination can be ATTRIBUTED rather than
guessed at.

audits/gate-never-asked/README.md carries the defect-class register and the
method for finding more: a signed scalar count, size, radix or index whose corpus
only ever takes the sensible range. A scan of the 280 landed changes for signed
scalar parameters produced ~40 candidates; 054-ultoa and 072-ultow were the two
with no negative literal anywhere, which is what started this.

+++ changes/073-ui64tow/reference.c
@@ -1,9 +1,18 @@
 // changes/073-ui64tow/reference.c — scalar oracle for _ui64tow.
+// THE RADIX IS CONVERTED WITH (unsigned), NOT (unsigned long long), AND THE DIFFERENCE IS REAL.
+// For every radix in 2..36 the two are identical, which is why the corpus -- which swept exactly
+// 2..36 -- could never tell them apart. Widening it to the out-of-range radices that ucrtbase
+// accepts made this model the ONLY one of the three that disagreed:
+//     v=4294967295, radix=-1:   ours '10'   ucrtbase '10'   this model 'V'
+// ucrtbase converts the radix to a 32-bit UNSIGNED value, so -1 is 4294967295 and
+// 4294967295/4294967295 = 1 remainder 0 -> "10". A 64-bit sign extension makes it
+// 0xFFFFFFFFFFFFFFFF instead, and a completely different answer. The implementation was right and
+// the oracle was wrong; its siblings 055 and 057 already used (unsigned).
 #include <wchar.h>
 wchar_t* ref_ui64tow(unsigned long long v, wchar_t* s, int radix){
     wchar_t tmp[70]; int n=0;
     if(v==0) tmp[n++]=L'0';
-    while(v){ unsigned d=(unsigned)(v%(unsigned long long)radix); tmp[n++]=(wchar_t)(d<10?L'0'+d:L'a'+d-10); v/=(unsigned long long)radix; }
+    while(v){ unsigned d=(unsigned)(v%(unsigned)radix); tmp[n++]=(wchar_t)(d<10?L'0'+d:L'a'+d-10); v/=(unsigned)radix; }
     for(int i=0;i<n;i++) s[i]=tmp[n-1-i];
     s[n]=0; return s;
 }
